backend/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python-headless>=4.5.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
